requirements.txt
# Quantum computing
qiskit>=1.3.0
qiskit-ibm-runtime>=0.34.0
qiskit-aer>=0.15.0
qiskit-machine-learning>=0.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
